refactor: extract TranscriptionRecovery presenters + support types
Bundles 2 domain-coherent extractions from TranscriptionRecoveryController.swift:

**Presenters** → TranscriptionRecoveryPresenters.swift:
- @MainActor final class RetryTranscriptionStatusPresenter
- @MainActor final class PendingTranscriptionRetryFailureHandler
- @MainActor final class RetryableSessionPreservationPresenter
- SHA-256: c19b4ba434b031267f5279052b652897499075f12c45103cf0cddc33828ddcee
  (lines 4-165 of origin/main:TranscriptionRecoveryController.swift)

**Support types** → TranscriptionRecoverySupport.swift:
- PendingTranscriptionRetryContext (struct)
- PendingTranscriptionRetryResolution (enum)
- PendingTranscriptionRetryFailure (struct)
- RetryableSessionPreservationResult (enum)
- SHA-256: d9895fcb211cf81c8b67a53f753ef50892d486b8cbc655227d7eed38bab5fc29
  (lines 439-461)

Applied memory rule "watch for annotations on line immediately preceding
deleted blocks": verified @MainActor at line 167 (for the controller class)
is preserved before deleting the presenter range 4-166.

Reduces TranscriptionRecoveryController.swift from 461 to 275 lines.

Tests: 667 (unchanged).
